# src/main/resources/openapi_napolya.yaml
openapi: 3.0.3
info:
  title: API na_polya 3.0
  version: 1.0.11
  description: Все необходимые API для сайта На_поля
...
paths:
...
  /fields/{id}:
    get:
      tags:
        - Field
      summary: Получить информацию о поле
      parameters:
        - name: id
          in: path
          required: true
          description: ID поля
          schema:
            type: integer
      responses:
        '200':
          description: Информация о поле
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        '404':
          description: Поле не найдено

    put:
      tags:
        - Field
      summary: Обновить информацию о поле
      parameters:
        - name: id
          in: path
          required: true
          description: ID поля
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Field'
      responses:
        '200':
          description: Поле успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        '404':
          description: Поле не найдено
  
    delete:
      tags:
        - Field
      summary: Удалить поле
      parameters:
        - name: id
          in: path
          required: true
          description: ID поля
          schema:
            type: integer
      responses:
        '204':
          description: Поле успешно удалено
        '404':
          description: Поле не найдено
  /games:
    get:
      tags:
        - Game
      summary: Получить список названий игр
      responses:
        '200':
          description: Успешный ответ с массивом названий игр
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: Список названий игр
                  example: ["Мега супер игра", "Кубок Полей"]
    post:
      tags:
        - Game
      summary: Создать новую игру
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
      responses:
        '201':
          description: Игра успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
  /games/{id}:
    get:
      tags:
        - Game
      summary: Получить информацию об игре
      parameters:
        - name: id
          in: path
          required: true
          description: ID игры
          schema:
            type: integer
      responses:
        '200':
          description: Информация об игре
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Игра не найдена

    put:
      tags:
        - Game
      summary: Обновить информацию об игре
      parameters:
        - name: id
          in: path
          required: true
          description: ID игры
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
      responses:
        '200':
          description: Игра успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Игра не найдена
    
    delete:
      tags:
        - Game
      summary: Удалить игру
      parameters:
        - name: id
          in: path
          required: true
          description: ID игры
          schema:
            type: integer
      responses:
        '204':
          description: Игра успешно удалена
        '404':
          description: Игра не найдена
...
components:
  schemas:
...
    Field:  # Определение сущности "Поле"
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор поля
          example: 1
        name:
          type: string
          description: Название поля
          example: "Арена Спорта"
        description:
          type: string
          description: Описание поля
          example: "Большой стадион с искусственным покрытием для футбольных матчей."
        location:
          type: string
          description: Местоположение поля
          example: "ул. Спортивная, д. 1, г. Санкт-Петербург"
        provider_id:
          type: integer
          description: id провайдера поля
          example: 2
        games_ids:
          type: array
          description: id игр, проводимых на этом поле
          items:
            type: integer
          example: [101, 102, 103]
      required:
        - id
        - name
        - location
        - provider_id
    Game: # Определение сущности "Игра"
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор игры
          example: 1
        name:
          type: string
          description: Название игры
          example: "Мега супер игра"
        description:
          type: string
          description: Описание игры
          example: "Друзья сыграют в дружеском матче в эту субботу."
        start_date:
          type: string
          description: Дата и время начала игры
          format: date-time
          example: "2024-11-16T15:00:00Z"
        status:
          type: string
          description: Статус игры
          enum: [ожидание, активная, завершена, запись закрыта]
          example: "ожидание"
        creator_id:
          type: integer
          description: id администратора, создавшего игру
          example: 2
        field_id:
          type: integer
          description: id поля, на котором будет проводиться игра
          example: 1
        teams_ids:
          type: array
          description: id команд, участвующих в игре
          items:
            type: integer
          example: [10, 11]
        queue_id:
          type: integer
          description: id очереди на игру
          example: 1

      required:
        - id
        - name
        - start_date
        - creator_id
        - field_id
...